구간 등록 단위테스트 작성
