+++ src/main/java/nextstep/subway/line/LineRepository.java
@@ -1,0 +1,7 @@
+package nextstep.subway.line;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import nextstep.subway.line.entity.Line;
+
+public interface LineRepository extends JpaRepository<Line, Long> {
+}
