+++ src/main/java/nextstep/subway/line/dto/LineCreateRequest.java
@@ -1,0 +1,42 @@
+package nextstep.subway.line.dto;
+
+public class LineCreateRequest {
+
+    private String name;
+
+    private String color;
+
+    private Long upStationId;
+
+    private Long downStationId;
+
+    private int distance;
+
+    public LineCreateRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
+        this.name = name;
+        this.color = color;
+        this.upStationId = upStationId;
+        this.downStationId = downStationId;
+        this.distance = distance;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+
+    public Long getUpStationId() {
+        return upStationId;
+    }
+
+    public Long getDownStationId() {
+        return downStationId;
+    }
+
+    public int getDistance() {
+        return distance;
+    }
+}
