+++ src/main/java/nextstep/subway/line/exception/LineException.java
@@ -1,0 +1,19 @@
+package nextstep.subway.line.exception;
+
+public class LineException extends RuntimeException {
+
+    public LineException() {
+    }
+
+    public LineException(String s) {
+        super(s);
+    }
+
+    public LineException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public LineException(Throwable cause) {
+        super(cause);
+    }
+}
