+++ src/main/java/nextstep/subway/station/Station.java
@@ -1,0 +1,29 @@
+package nextstep.subway.station;
+
+import javax.persistence.*;
+
+@Entity
+public class Station {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(length = 20, nullable = false)
+    private String name;
+
+    public Station() {
+    }
+
+    public Station(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
