+++ src/main/java/nextstep/subway/station/StationRepository.java
@@ -1,0 +1,6 @@
+package nextstep.subway.station;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StationRepository extends JpaRepository<Station, Long> {
+}
