+++ src/main/java/nextstep/subway/station/dto/StationRequest.java
@@ -1,0 +1,9 @@
+package nextstep.subway.station.dto;
+
+public class StationRequest {
+    private String name;
+
+    public String getName() {
+        return name;
+    }
+}
