+++ src/main/java/nextstep/subway/station/service/StationDataService.java
@@ -1,0 +1,24 @@
+package nextstep.subway.station.service;
+
+import org.springframework.stereotype.Service;
+import nextstep.subway.station.Station;
+import nextstep.subway.station.exception.StationException;
+import nextstep.subway.station.StationRepository;
+
+import javax.transaction.Transactional;
+
+@Transactional
+@Service
+public class StationDataService {
+
+    private final StationRepository stationRepository;
+
+    public StationDataService(StationRepository stationRepository) {
+        this.stationRepository = stationRepository;
+    }
+
+    public Station findStation(Long id) {
+        return stationRepository.findById(id)
+                .orElseThrow(() -> new StationException("존재하지 않는 역입니다."));
+    }
+}
