+++ src/main/java/nextstep/subway/station/service/StationService.java
@@ -1,0 +1,42 @@
+package nextstep.subway.station.service;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import nextstep.subway.station.Station;
+import nextstep.subway.station.StationRepository;
+import nextstep.subway.station.dto.StationResponse;
+import nextstep.subway.station.dto.StationRequest;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+@Transactional(readOnly = true)
+public class StationService {
+    private final StationRepository stationRepository;
+
+    public StationService(StationRepository stationRepository) {
+        this.stationRepository = stationRepository;
+    }
+
+    @Transactional
+    public StationResponse saveStation(StationRequest stationRequest) {
+        Station station = stationRepository.save(new Station(stationRequest.getName()));
+        return createStationResponse(station);
+    }
+
+    public List<StationResponse> findAllStations() {
+        return stationRepository.findAll().stream()
+                .map(this::createStationResponse)
+                .collect(Collectors.toList());
+    }
+
+    @Transactional
+    public void deleteStationById(Long id) {
+        stationRepository.deleteById(id);
+    }
+
+    private StationResponse createStationResponse(Station station) {
+        return StationResponse.ofEntity(station);
+    }
+}
