--- src/main/java/subway/line/dto/LineUpdateRequest.java
@@ -1,21 +1,0 @@
-package subway.line.dto;
-
-public class LineUpdateRequest {
-
-    private String name;
-
-    private String color;
-
-    public LineUpdateRequest(String name, String color) {
-        this.name = name;
-        this.color = color;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-}
