--- src/main/java/subway/line/exception/LineException.java
@@ -1,19 +1,0 @@
-package subway.line.exception;
-
-public class LineException extends RuntimeException {
-
-    public LineException() {
-    }
-
-    public LineException(String s) {
-        super(s);
-    }
-
-    public LineException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public LineException(Throwable cause) {
-        super(cause);
-    }
-}
