--- src/main/java/subway/section/Section.java
@@ -1,58 +1,0 @@
-package subway.section;
-
-import subway.line.entity.Line;
-import subway.station.Station;
-
-import javax.persistence.*;
-
-@Entity
-public class Section {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private int distance;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "up_station_id", foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
-    private Station upStation;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "down_station_id", foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
-    private Station downStation;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "line_id", foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
-    private Line line;
-
-    public Section() {
-    }
-
-    public Section(int distance, Station upStation, Station downStation, Line line) {
-        this.distance = distance;
-        this.upStation = upStation;
-        this.downStation = downStation;
-        this.line = line;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public int getDistance() {
-        return distance;
-    }
-
-    public Station getUpStation() {
-        return upStation;
-    }
-
-    public Station getDownStation() {
-        return downStation;
-    }
-
-    public Line getLine() {
-        return line;
-    }
-}
