--- src/main/java/subway/station/Station.java
@@ -1,29 +1,0 @@
-package subway.station;
-
-import javax.persistence.*;
-
-@Entity
-public class Station {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(length = 20, nullable = false)
-    private String name;
-
-    public Station() {
-    }
-
-    public Station(String name) {
-        this.name = name;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
