--- src/main/java/subway/station/StationRepository.java
@@ -1,6 +1,0 @@
-package subway.station;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StationRepository extends JpaRepository<Station, Long> {
-}
