--- src/main/java/subway/station/dto/StationRequest.java
@@ -1,9 +1,0 @@
-package subway.station.dto;
-
-public class StationRequest {
-    private String name;
-
-    public String getName() {
-        return name;
-    }
-}
